ISSUE #2754 src for containers/federations were moved but not the tests
